examples/features/torch_batch_process/batch_inference/generate_embedding/requirements.txt
accelerate>=0.12.0
transformers>=4.28.1,<4.29.0
tokenizers>=0.13.3
datasets
pinecone-client
